Improve previous test code
