+++ GenericsAnalyzer/GenericsAnalyzer.Test/BaseDiagnosticTests.cs
@@ -6,9 +6,29 @@ using static GenericsAnalyzer.Test.Helpers.UsingsHelpers;
 
 namespace GenericsAnalyzer.Test
 {
+    public abstract class BaseDiagnosticTests<TAnalyzer> : BaseDiagnosticTests
+        where TAnalyzer : DiagnosticAnalyzer, new()
+    {
+        protected sealed override DiagnosticAnalyzer GetNewDiagnosticAnalyzerInstance() => new TAnalyzer();
+    }
+
     public abstract class BaseDiagnosticTests : IAnalyzerTestFixture
     {
-        public abstract DiagnosticDescriptor TestedDiagnosticRule { get; }
+        private DiagnosticDescriptor testedDiagnosticRule;
+
+        public virtual DiagnosticDescriptor TestedDiagnosticRule
+        {
+            get
+            {
+                if (testedDiagnosticRule != null)
+                    return testedDiagnosticRule;
+
+                // TODO: This will need major refactoring if a new diagnostic group will be introduced
+                var thisType = GetType();
+                var ruleID = thisType.Name.Substring(0, "GA0000".Length);
+                return testedDiagnosticRule = DiagnosticDescriptors.GetDiagnosticDescriptor(ruleID);
+            }
+        }
 
         protected ExpectedDiagnostic ExpectedDiagnostic => ExpectedDiagnostic.Create(TestedDiagnosticRule);
 
@@ -29,6 +49,18 @@ namespace GenericsAnalyzer.Test
         protected void AssertDiagnosticsWithUsings(string testCode)
         {
             AssertDiagnostics(WithUsings(testCode));
+        }
+
+        protected void AssertOrValidate(string testCode, bool assert)
+        {
+            if (assert)
+                AssertDiagnostics(testCode);
+            else
+                ValidateCode(testCode.Replace("↓", ""));
+        }
+        protected void AssertOrValidateWithUsings(string testCode, bool assert)
+        {
+            AssertOrValidate(WithUsings(testCode), assert);
         }
 
         [TestMethod]
